+++ swarm-derive/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.34.0 - unreleased
 
+- Adapt to interface changes in `libp2p-swarm`.
+  See [PR 4706](https://github.com/libp2p/rust-libp2p/pull/4076).
 
 ## 0.33.0